huggingface-sagemaker/requirements.txt
aws-cdk-lib==2.185.0
constructs>=10.0.0,<11.0.0
huggingface-hub==0.30.1
transformers==4.50.3
boto3